Combine detector and template generator, add labels
This combines the detector logic and the homography tools into a single
class ArUcoHomograpy, providing homograpy tools. That includes the
template generator, and the marker detection, but will later provide the
transformation coefficients.

This also changes the detection so that QML gets all of the corner
points of the markers in numeric form instead of just a string, and that
is used to add labels to the image showing which marker ID is detected
where.