Add support for multiple templates
This adds support for multiple templates - a Letter one and a Postcard
one at the moment -- this is so you can print a smaller template.

This is for later when I want to make a 3d printed template.

This removes the ability to scroll through all of the marker Ids.
I might want to return that capability at some point, because viewing
all of the different markers is nice.